requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-allauth==0.49.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-summernote==0.8.20.0
oauthlib==3.2.0
Pillow==9.0.1
PyJWT==2.3.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
